Readme.md: Explain how to configure javac options
For many users (myself included), it is not obvious how to configure
the classpath that the linter should use, and guidance for how to
do this is spread across open and closed issues.

This adds a short explanation to the Readme that shows how to pass
arguments to `javac`, and points out the pitfall that options
and their values need to be stored in separate array elements.